breaking up cmd and gocron packages. compiles. still a work in progress, this commit has not been tested
