--- cmd/const.go
@@ -1,7 +1,0 @@
-package cmd
-
-
-const VERSION string = "4.0.3"
-
-const missedJobs = "SELECT * FROM gocron WHERE (extract(epoch from now()) - lastruntime) > frequency;"
-const revivedJobs = "SELECT * FROM gocron WHERE alerted = true AND (extract(epoch from now()) - lastruntime) < frequency;"
