+++ libgocron/variables.go
@@ -1,12 +1,17 @@
-package cmd
+package libgocron
+
+const VERSION string = "5.0"
 
 
-var cfgFile      string
-var frontendPort string
-var summary      bool
-var verbose      bool
+//var cfgFile      string
+//var frontendPort string
+//var summary      bool
+//var verbose      bool = false
 
 var config Config
+
+
+
 
 
 type Cron struct {
